Create a proxy that intercepts native docker build calls
Begin work towards a per node proxy that can authenticate and authorize
specific docker build operations, removing the build pod from having
privileged access to a node. By emulating the docker api we ensure that
we can support existing clients while adding a framework for providing
additional backends that may not rely on a docker socket.

This uses a variant of the weave proxy to provide a simple intercepting
proxy to docker that can be layered with other logic.